受付APIのサービスを追加した

+++ Vagrantfile
@@ -20,29 +20,34 @@ Vagrant.configure(2) do |config|
     curl -o /etc/apt/sources.list.d/jma-receipt-xenial50.list https://ftp.orca.med.or.jp/pub/ubuntu/jma-receipt-xenial50.list
     curl https://ftp.orca.med.or.jp/pub/ubuntu/archive.key | apt-key add -
     apt-get update
+    apt-get upgrade -y
     apt-get install -y \
     curl \
-    jma-receipt \
     language-pack-ja \
     nginx \
     rsyslog \
     syslinux-common \
     uuid-runtime
-    apt-get upgrade -y
     update-locale LANG=ja_JP.UTF-8
     timedatectl set-timezone Asia/Tokyo
+    apt-get install -y jma-receipt
     echo "listen_addresses = '*'" >> /etc/postgresql/9.5/main/postgresql.conf
     echo 'host all all 0.0.0.0/0 trust' >> /etc/postgresql/9.5/main/pg_hba.conf
     echo 'DBENCODING="UTF-8"' >> /etc/jma-receipt/db.conf
-    echo "ormaster:$(md5pass ormaster):" > /etc/jma-receipt/passwd
-    chown orca:orca /etc/jma-receipt/passwd
-    chmod 644 /etc/jma-receipt/passwd
-    jma-setup
-    su orca - -c '/usr/lib/jma-receipt/bin/passwd_store.sh | nkf -w'
-    service postgresql restart
     jma-setup
     service jma-receipt start
-    su orca - -c '/usr/lib/jma-receipt/scripts/tools/run_master_upgrade.sh | nkf -w'
-    su orca - -c '/usr/lib/jma-receipt/bin/jma-receipt-program-upgrade.sh | nkf -w'
+    if [ ! -f /etc/jma-receipt/passwd ]; then
+      echo "ormaster:$(md5pass ormaster):" > /etc/jma-receipt/passwd
+      chown orca:orca /etc/jma-receipt/passwd
+      chmod 644 /etc/jma-receipt/passwd
+      su orca - -c '/usr/lib/jma-receipt/bin/passwd_store.sh | nkf -w'
+    fi
+    # jma-setup
+    # service jma-receipt restart
+    # su orca - -c '/usr/lib/jma-receipt/scripts/tools/run_master_upgrade.sh | nkf -w'
+    # su orca - -c '/usr/lib/jma-receipt/bin/jma-receipt-program-upgrade.sh | nkf -w'
+    # service jma-receipt stop
+    # service postgresql restart
+    # service jma-receipt start
   SHELL
 end
